[rabbitmq] Run as rabbitmq and avoid restart (rabbitmq 3.8 compatible)
Setting the fsGroup and runAsGroup should ensure that
the data is readable by non-root users.
Kubernetes can make use of the kernel to remap users, but may fall
back to a chown as done in the script if that isn't possible

Since we are already running everything then as rabbitmq-user,
we also can avoid starting the server up twice

The startup-script was also removed in some version after 3.6.
The new version works at least with 3.8